requirements.txt
slack_bolt
openai
python-dotenv
Pillow
requests
openpyxl 